+++ backend/canisters/user_index/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Push all users to OpenStorage with the new storage limits ([#3104](https://github.com/open-ic/open-chat/pull/3104))
+
 ### Removed
 
 - Removed code only needed for the previous upgrade ([#3102](https://github.com/open-ic/open-chat/pull/3102))